Begin migrate tests
