+++ Cargo.toml
@@ -39,7 +39,7 @@ serde_json = "1"
 
 [dev-dependencies]
 serial_test = "3"
-tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+tokio = { version = "1", features = ["macros", "rt", "time", "test-util"] }
 actix-rt = "2"
 actix-web = "4"
 actix-http = "3.9.0"
