+++ tests/mod.rs
@@ -1,0 +1,3 @@
+mod tests_token;
+mod tests_stats;
+mod tests_network;
