+++ tests/tests_stats/mod.rs
@@ -1,0 +1,2 @@
+mod tokencount;
+mod stats;
