+++ tests/tests_token/mod.rs
@@ -1,0 +1,3 @@
+mod auth;
+mod crypto;
+mod csrf;
